Added the leave command
Which I couldn't test because discord is a group of LOVELY AND AMAZING
people. I can not go on for a single day without complaining
about how fucking unstable their API is, though.
Sometimes it works, sometimes it just doesn't. And you know what?
My bad for not trying to send one request 34 times in hopes that
at least one of them will not return an error.

You know what? I think it was a great idea to force an unstable API
upon their users. Slash (Application) commands will be the only supported
way to control bots and that is fine, I guess. But the problem is that
the API is -- to this day -- unstable. ????????????????????

Discord devs, with my uttermost respect: suck my dick.